compiler options and codegen split
