+++ cli-rs/src/main.rs
@@ -7,8 +7,9 @@ use env_logger::fmt::Color;
 use env_logger::{Builder, Target};
 use log::LevelFilter;
 
-use powdr_number::{BigUint, Bn254Field, FieldElement, GoldilocksField};
+use powdr_number::{BabyBearField, BigUint, Bn254Field, FieldElement, GoldilocksField, KnownField};
 use powdr_pipeline::Pipeline;
+use powdr_riscv::{CompilerOptions, RuntimeLibs};
 use powdr_riscv_executor::ProfilerOptions;
 
 use std::ffi::OsStr;
@@ -20,10 +21,22 @@ use strum::{Display, EnumString, EnumVariantNames};
 
 #[derive(Clone, EnumString, EnumVariantNames, Display)]
 pub enum FieldArgument {
+    #[strum(serialize = "bb")]
+    Bb,
     #[strum(serialize = "gl")]
     Gl,
     #[strum(serialize = "bn254")]
     Bn254,
+}
+
+impl FieldArgument {
+    pub fn as_known_field(&self) -> KnownField {
+        match self {
+            FieldArgument::Bb => KnownField::BabyBearField,
+            FieldArgument::Gl => KnownField::GoldilocksField,
+            FieldArgument::Bn254 => KnownField::Bn254Field,
+        }
+    }
 }
 
 #[derive(Parser)]
@@ -196,28 +209,26 @@ fn run_command(command: Commands) {
             output_directory,
             coprocessors,
             continuations,
-        } => {
-            call_with_field!(compile_rust::<field>(
-                &file,
-                Path::new(&output_directory),
-                coprocessors,
-                continuations
-            ))
-        }
+        } => compile_rust(
+            &file,
+            field.as_known_field(),
+            Path::new(&output_directory),
+            coprocessors,
+            continuations,
+        ),
         Commands::RiscvElf {
             file,
             field,
             output_directory,
             coprocessors,
             continuations,
-        } => {
-            call_with_field!(compile_riscv_elf::<field>(
-                &file,
-                Path::new(&output_directory),
-                coprocessors,
-                continuations
-            ))
-        }
+        } => compile_riscv_elf(
+            &file,
+            field.as_known_field(),
+            Path::new(&output_directory),
+            coprocessors,
+            continuations,
+        ),
         Commands::Execute {
             file,
             field,
@@ -243,6 +254,7 @@ fn run_command(command: Commands) {
             };
             call_with_field!(execute::<field>(
                 Path::new(&file),
+                field.as_known_field(),
                 split_inputs(&inputs),
                 Path::new(&output_directory),
                 continuations,
@@ -260,56 +272,32 @@ fn run_command(command: Commands) {
 }
 
 #[allow(clippy::too_many_arguments)]
-fn compile_rust<F: FieldElement>(
+fn compile_rust(
     file_name: &str,
+    field: KnownField,
     output_dir: &Path,
     coprocessors: Option<String>,
     continuations: bool,
 ) -> Result<(), Vec<String>> {
-    let mut runtime = match coprocessors {
-        Some(list) => {
-            powdr_riscv::Runtime::try_from(list.split(',').collect::<Vec<_>>().as_ref()).unwrap()
-        }
-        None => powdr_riscv::Runtime::base(),
-    };
-
-    if continuations {
-        if runtime.has_submachine("poseidon_gl") {
-            return Err(vec![
-                "Poseidon continuations mode is chosen automatically and incompatible with the chosen standard Poseidon coprocessor".to_string(),
-            ]);
-        }
-        runtime = runtime.with_poseidon_for_continuations();
-    }
-
-    powdr_riscv::compile_rust::<F>(file_name, output_dir, true, &runtime, continuations, None)
+    let libs = coprocessors_to_options(coprocessors)?;
+    let options = CompilerOptions::new(field, libs, continuations);
+    powdr_riscv::compile_rust(file_name, options, output_dir, true, None)
         .ok_or_else(|| vec!["could not compile rust".to_string()])?;
 
     Ok(())
 }
 
-fn compile_riscv_elf<F: FieldElement>(
+fn compile_riscv_elf(
     input_file: &str,
+    field: KnownField,
     output_dir: &Path,
     coprocessors: Option<String>,
     continuations: bool,
 ) -> Result<(), Vec<String>> {
-    let runtime = match coprocessors {
-        Some(list) => {
-            powdr_riscv::Runtime::try_from(list.split(',').collect::<Vec<_>>().as_ref()).unwrap()
-        }
-        None => powdr_riscv::Runtime::base(),
-    };
-
-    powdr_riscv::compile_riscv_elf::<F>(
-        input_file,
-        Path::new(input_file),
-        output_dir,
-        true,
-        &runtime,
-        continuations,
-    )
-    .ok_or_else(|| vec!["could not translate RISC-V executable".to_string()])?;
+    let libs = coprocessors_to_options(coprocessors)?;
+    let options = CompilerOptions::new(field, libs, continuations);
+    powdr_riscv::compile_riscv_elf(input_file, Path::new(input_file), options, output_dir, true)
+        .ok_or_else(|| vec!["could not translate RISC-V executable".to_string()])?;
 
     Ok(())
 }
@@ -317,6 +305,7 @@ fn compile_riscv_elf<F: FieldElement>(
 #[allow(clippy::too_many_arguments)]
 fn execute<F: FieldElement>(
     file_name: &Path,
+    field: KnownField,
     inputs: Vec<F>,
     output_dir: &Path,
     continuations: bool,
@@ -335,7 +324,7 @@ fn execute<F: FieldElement>(
 
     match (witness, continuations) {
         (false, true) => {
-            powdr_riscv::continuations::rust_continuations_dry_run(&mut pipeline, profiling);
+            powdr_riscv::continuations::rust_continuations_dry_run(&mut pipeline, field, profiling);
         }
         (false, false) => {
             let program = pipeline.compute_asm_string().unwrap().clone();
@@ -350,8 +339,11 @@ fn execute<F: FieldElement>(
             log::info!("Execution trace length: {}", trace.len);
         }
         (true, true) => {
-            let dry_run =
-                powdr_riscv::continuations::rust_continuations_dry_run(&mut pipeline, profiling);
+            let dry_run = powdr_riscv::continuations::rust_continuations_dry_run(
+                &mut pipeline,
+                field,
+                profiling,
+            );
             powdr_riscv::continuations::rust_continuations(pipeline, generate_witness, dry_run)?;
         }
         (true, false) => {
@@ -360,4 +352,20 @@ fn execute<F: FieldElement>(
     }
 
     Ok(())
+}
+
+fn coprocessors_to_options(coprocessors: Option<String>) -> Result<RuntimeLibs, Vec<String>> {
+    let mut libs = RuntimeLibs::new();
+    if let Some(list) = coprocessors {
+        let names = list.split(',').collect::<Vec<_>>();
+        for name in names {
+            match name {
+                "poseidon_gl" => libs = libs.with_poseidon(),
+                "keccakf" => libs = libs.with_keccak(),
+                "arith" => libs = libs.with_arith(),
+                _ => return Err(vec![format!("Invalid co-processor specified: {name}")]),
+            }
+        }
+    }
+    Ok(libs)
 }
