+++ cli-rs/src/util.rs
@@ -13,6 +13,7 @@ macro_rules! clap_enum_variants {
 macro_rules! call_with_field {
     ($function:ident::<$field:ident>($($args:expr),*) ) => {
         match $field {
+            FieldArgument::Bb => $function::<BabyBearField>($($args),*),
             FieldArgument::Gl => $function::<GoldilocksField>($($args),*),
             FieldArgument::Bn254 => $function::<Bn254Field>($($args),*),
         }
