+++ pipeline/src/test_util.rs
@@ -57,21 +57,21 @@ pub fn make_prepared_pipeline<T: FieldElement>(
     pipeline
 }
 
-/// Tests witness generation, pilcom, halo2 and estark.
-/// Does NOT test plonky3.
+/// Tests witness generation, pilcom, halo2, estark and plonky3.
 pub fn regular_test(file_name: &str, inputs: &[i32]) {
     let inputs_gl = inputs.iter().map(|x| GoldilocksField::from(*x)).collect();
     let pipeline_gl = make_prepared_pipeline(file_name, inputs_gl, vec![]);
     test_pilcom(pipeline_gl.clone());
-    gen_estark_proof(pipeline_gl);
+    gen_estark_proof(pipeline_gl.clone());
+    test_plonky3_pipeline(pipeline_gl);
 
     let inputs_bn = inputs.iter().map(|x| Bn254Field::from(*x)).collect();
     let pipeline_bn = make_prepared_pipeline(file_name, inputs_bn, vec![]);
     test_halo2(pipeline_bn);
 
     let inputs_bb = inputs.iter().map(|x| BabyBearField::from(*x)).collect();
-    let mut pipeline_bb = make_prepared_pipeline(file_name, inputs_bb, vec![]);
-    pipeline_bb.compute_witness().unwrap();
+    let pipeline_bb = make_prepared_pipeline(file_name, inputs_bb, vec![]);
+    test_plonky3_pipeline(pipeline_bb);
 }
 
 pub fn regular_test_without_babybear(file_name: &str, inputs: &[i32]) {
@@ -301,6 +301,45 @@ pub fn test_plonky3_with_backend_variant<T: FieldElement>(
         .from_file(resolve_test_file(file_name))
         .with_prover_inputs(inputs)
         .with_backend(backend, None);
+
+    // Generate a proof
+    let proof = pipeline.compute_proof().cloned().unwrap();
+
+    let publics: Vec<T> = pipeline
+        .publics()
+        .clone()
+        .unwrap()
+        .iter()
+        .map(|(_name, v)| *v)
+        .collect();
+
+    pipeline.verify(&proof, &[publics.clone()]).unwrap();
+
+    if pipeline.optimized_pil().unwrap().constant_count() > 0 {
+        // Export verification Key
+        let output_dir = pipeline.output_dir().as_ref().unwrap();
+        let vkey_file_path = output_dir.join("verification_key.bin");
+        buffered_write_file(&vkey_file_path, |writer| {
+            pipeline.export_verification_key(writer).unwrap()
+        })
+        .unwrap();
+
+        let mut pipeline = pipeline.with_vkey_file(Some(vkey_file_path));
+
+        // Verify the proof again
+        pipeline.verify(&proof, &[publics]).unwrap();
+    }
+}
+
+#[cfg(feature = "plonky3")]
+pub fn test_plonky3_pipeline<T: FieldElement>(pipeline: Pipeline<T>) {
+    let mut pipeline = pipeline.with_backend(powdr_backend::BackendType::Plonky3Composite, None);
+
+    pipeline.compute_witness().unwrap();
+
+    if !should_generate_proofs() {
+        return;
+    }
 
     // Generate a proof
     let proof = pipeline.compute_proof().cloned().unwrap();
